requirements.txt
fastapi==0.95.1
sqlalchemy==2.0.15
pydantic==1.10.7
mysql-connector-python==8.0.33
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi-cache2==0.2.1
pymysql==1.1.1
email-validator==2.1.1
uvicorn==0.29.0